--- output/1.txt
@@ -1,1 +1,0 @@
-1111
